refactor(discovery): separate discovery allowlist logic into a new module
- Moved functions related to manifest handling from `discovery-allowlist.ts` to a new file `discovery-allowlist-manifest.ts`, improving code organization.
- Updated imports across various files to reflect the new structure, ensuring that the application continues to function correctly with the refactored code.
- Enhanced `.env.example` to clarify DMZ host binding options for production environments.
