+++ .env.example
@@ -51,7 +51,8 @@ DATABASE_URL=postgresql://user:password@ep-example.us-east-2.aws.neon.tech/neond
 # Host publish binds to 127.0.0.1 by default (SIGNER_DMZ_BIND_HOST). Use
 # 127.0.0.1 in URLs below when PymtHouse runs on the same host; remote callers
 # need your server's hostname/IP and valid DMZ JWTs (scope=sign:job / admin).
-# SIGNER_DMZ_BIND_HOST=0.0.0.0
+# Public TLS: host Apache → 127.0.0.1:8080 (see deploy/host-apache/). Do not bind 0.0.0.0 in production.
+# SIGNER_DMZ_BIND_HOST=127.0.0.1
 # SIGNER_DMZ_HOST_PORT=8080
 SIGNER_INTERNAL_URL=http://127.0.0.1:8080
 # Apache CLI admin endpoint (admin scope): path-mounted on the same host/port as
