+++ src/app/api/v1/apps/[id]/manifest/route.test.ts
@@ -5,9 +5,11 @@ import { eq } from "drizzle-orm";
 import { db } from "@/db/index";
 import { planCapabilityBundles, plans } from "@/db/schema";
 import {
+} from "@/lib/discovery-allowlist";
+import {
   ALLOW_ALL_MANIFEST_ETAG,
   ALLOW_ALL_MANIFEST_RESPONSE,
-} from "@/lib/discovery-allowlist";
+} from "@/lib/discovery-allowlist-manifest";
 import { run } from "@/test-utils/db-guard";
 import { cleanupTestApp, seedDeveloperAppWithClient } from "@/test-utils/fixtures";
 import {
